+++ js/index.js
@@ -15,7 +15,8 @@
       './prev-next-button',
       './page-dots',
       './player',
-      './add-remove-cell'
+      './add-remove-cell',
+      './lazyload'
     ], factory );
   } else if ( typeof exports == 'object' ) {
     // CommonJS
@@ -25,7 +26,8 @@
       require('./prev-next-button'),
       require('./page-dots'),
       require('./player'),
-      require('./add-remove-cell')
+      require('./add-remove-cell'),
+      require('./lazyload')
     );
   }